+++ framework/ui/src/TextField.cpp
@@ -31,6 +31,8 @@ namespace bdn::ui
             core->autocorrectionType.bind(autocorrectionType);
             core->returnKeyType.bind(returnKeyType);
             core->placeholder.bind(placeholder);
+            core->textInputType.bind(textInputType);
+            core->obscureInput.bind(obscureInput);
 
             _submitCallbackReceiver = core->submitCallback.set([=]() {
                 SubmitEvent evt(shared_from_this());